requirements.txt
alembic==1.0.8
Babel==2.6.0
blinker==1.4
chardet==3.0.4
Click==7.0
cymysql==0.9.1
dnspython==1.16.0
esdk-obs-python==3.0.5
etcd3==0.10.0
Flask==1.0.1
Flask-Admin==1.5.3
Flask-BabelEx==0.9.3
Flask-Cache==0.13.1
Flask-Cors==2.1.0
Flask-HTTPAuth==2.7.0
Flask-Login==0.4.1
Flask-Mail==0.9.1
Flask-Migrate==2.2.1
Flask-Script==2.0.6
Flask-SQLAlchemy==2.3.2
Flask-WTF==0.14.2
grpcio==1.20.1
idna==2.7
itsdangerous==1.1.0
Jinja2==2.10
lxml==4.2.5
Mako==1.0.8
MarkupSafe==1.1.1
protobuf==3.7.1
pymongo==3.8.0
PyMySQL==0.7.11
python-dateutil==2.8.0
python-editor==1.0.4
python-etcd==0.4.5
pytz==2019.1
requests==2.19.1
six==1.12.0
speaklater==1.3
SQLAlchemy==1.3.1
tenacity==5.0.4
urllib3==1.23
Werkzeug==0.15.2
WTForms==2.2.1